+++ Configuration/ESBuild.d.ts
@@ -1,0 +1,8 @@
+import type { BuildOptions } from "esbuild";
+
+/**
+ * @module ESBuild
+ *
+ */
+declare const _default: BuildOptions;
+export default _default;