+++ Source/Class/Delete.d.ts
@@ -1,7 +1,0 @@
-#!/usr/bin/env node
-/**
- * @module Delete
- *
- */
-declare const _default: import("commander").Command;
-export default _default;