+++ Source/Function/Project.d.ts
@@ -1,7 +1,0 @@
-/**
- * @module Project
- *
- * Lists all Cloudflare Pages project names for an account.
- */
-declare const _default: (AccountID: string, Headers: HeadersInit) => Promise<string[]>;
-export default _default;