+++ Source/Interface/Environment.d.ts
@@ -1,6 +1,0 @@
-/**
- * @module Environment
- *
- */
-export default interface Interface {
-}