+++ Source/Interface/Response.d.ts
@@ -1,8 +1,0 @@
-import type { Response } from "@cloudflare/workers-types/experimental/index.js";
-/**
- * @module Response
- *
- */
-export default interface Interface {
-    (Message?: unknown, Status?: number): Promise<Response>;
-}