+++ Target/Class/Delete.d.js
@@ -1,2 +1,0 @@
-#!/usr/bin/env node
-var e=_default;export{e as default};