+++ Target/Function/Deployment.d.ts
@@ -3,5 +3,9 @@
  *
  * Lists all deployment IDs for a Cloudflare Pages project, handling pagination.
  */
-declare const _default: (AccountID: string, Project: string, Headers: HeadersInit) => Promise<string[]>;
+declare const _default: (
+	AccountID: string,
+	Project: string,
+	Headers: HeadersInit,
+) => Promise<string[]>;
 export default _default;