+++ Target/Function/Project.d.ts
@@ -3,5 +3,8 @@
  *
  * Lists all Cloudflare Pages project names for an account.
  */
-declare const _default: (AccountID: string, Headers: HeadersInit) => Promise<string[]>;
+declare const _default: (
+	AccountID: string,
+	Headers: HeadersInit,
+) => Promise<string[]>;
 export default _default;