+++ Target/Function/Response.js
@@ -1,1 +1,6 @@
-var t=(async(e=null,n=200)=>new Response(JSON.stringify(e),{status:n,headers:{"Content-Type":"application/json;charset=utf-8"}}));export{t as default};
+var t = async (e = null, n = 200) =>
+	new Response(JSON.stringify(e), {
+		status: n,
+		headers: { "Content-Type": "application/json;charset=utf-8" },
+	});
+export { t as default };