+++ Target/Function/Worker.d.js
@@ -1,1 +1,0 @@
-var e=_default;export{e as default};