+++ Target/Variable/Environment.d.js
@@ -1,1 +1,0 @@
-var o=_default;export{o as default};