+++ Target/Variable/Environment.d.ts
@@ -1,9 +1,22 @@
 export declare const string: typeof import("zod").string;
-declare const _default: import("zod").ZodObject<{
-    Email: import("zod").ZodDefault<import("zod").ZodOptional<import("zod").ZodString>>;
-    ID: import("zod").ZodDefault<import("zod").ZodOptional<import("zod").ZodString>>;
-    Key: import("zod").ZodDefault<import("zod").ZodOptional<import("zod").ZodString>>;
-    Token: import("zod").ZodDefault<import("zod").ZodOptional<import("zod").ZodString>>;
-    Project: import("zod").ZodDefault<import("zod").ZodOptional<import("zod").ZodString>>;
-}, import("zod/v4/core").$strip>;
+declare const _default: import("zod").ZodObject<
+	{
+		Email: import("zod").ZodDefault<
+			import("zod").ZodOptional<import("zod").ZodString>
+		>;
+		ID: import("zod").ZodDefault<
+			import("zod").ZodOptional<import("zod").ZodString>
+		>;
+		Key: import("zod").ZodDefault<
+			import("zod").ZodOptional<import("zod").ZodString>
+		>;
+		Token: import("zod").ZodDefault<
+			import("zod").ZodOptional<import("zod").ZodString>
+		>;
+		Project: import("zod").ZodDefault<
+			import("zod").ZodOptional<import("zod").ZodString>
+		>;
+	},
+	import("zod/v4/core").$strip
+>;
 export default _default;